Add Color Border
This will add an active color border around the color-picker.

Issues
===

Focus is being selected after refresh of popup, however, when you close and re-open
focus is still brought to first element.

There are two trains of thought here I believe:

1. Accessibility for keyboard users would be correct by placing focus on first element
1. Design is wrong because focus is so bold, it appears to indicate selected color.

Now since the above improvement is handling active color, the question is if focus
is indeed to bold, and we need to scale back, or if this implementation is fine.