Change Content Script Matching Glob Pattern
Since google doesn't make it easy to match all of their sites. I used a list
of the top 10 user data searches and added those countries. Hopefully that hits
a broad range of users, and I can go back in and add countries as needed.

I hope this becomes easier in the future.