Update UI And Add Toggle Functionality
This is the bulk of the whitelist rewrite. It creates the ability for
users to control what sites they want color-links to be on and which
they don't.

This pr also introduces a new color palette along with refresh
capabilities. Previously users had to refresh their page to see
results, this (may annoyingly) will refresh for them and therefore
enable the stylesheet or disable it. This functionality is primarily
needed for when a user disables color links on a page they previously
had it enabled.